cvat/requirements/base.txt
attrs==21.2.0
click==7.1.2
Django==3.2.12
django-appconf==1.0.4
django-auth-ldap==2.2.0
django-compressor==2.4
django-rq==2.3.2
EasyProcess==0.3
Pillow==9.0.1
numpy==1.22.0
python-ldap==3.4.0
pytz==2020.1
pyunpack==0.2.1
rcssmin==1.0.6
redis==3.5.3
rjsmin==1.1.0
requests==2.26.0
rq==1.5.1
rq-scheduler==0.10.0
sqlparse==0.4.2
django-sendfile2==0.6.1
dj-pagination==2.5.0
python-logstash-async==2.2.0
rules==2.2
GitPython==3.1.27
coreapi==2.3.3
django-filter==2.4.0
Markdown==3.2.2
djangorestframework==3.12.4
Pygments==2.7.4
drf-spectacular==0.21.2
Shapely==1.7.1
pdf2image==1.14.0
django-rest-auth[with_social]==0.9.5
opencv-python-headless==4.5.5.62
h5py==3.6.0
django-cors-headers==3.5.0
furl==2.1.0
av==8.0.2 --no-binary=av
tensorflow==2.8.0 # Optional requirement of Datumaro. Use tensorflow-macos==2.8.0 for Mac M1
# The package is used by pyunpack as a command line tool to support multiple
# archives. Don't use as a python module because it has GPL license.
patool==1.12
diskcache==5.0.2
open3d==0.14.1
boto3==1.17.61
azure-storage-blob==12.8.1
google-cloud-storage==1.42.0
# --no-binary=datumaro: workaround for pip to install
# opencv-headless instead of regular opencv, to actually run setup script
datumaro==0.2.0 --no-binary=datumaro
urllib3>=1.26.5 # not directly required, pinned by Snyk to avoid a vulnerability
natsort==8.0.0
mistune>=2.0.1 # not directly required, pinned by Snyk to avoid a vulnerability
dnspython==2.2.0
tk-tools